fix: downcast version numbers to unsigned int

+++ src/http/connection.hpp
@@ -116,11 +116,12 @@ private:
     {
         logger::debug([this] {
             const char *target = request_.target().data();
-            return fmt::format("Received \"{} {} HTTP/{}.{}\"",
-                               request_.method_string(),
-                               target,
-                               request_.version() / 10,
-                               request_.version() % 10);
+            return fmt::format(
+                "Received \"{} {} HTTP/{}.{}\"",
+                request_.method_string(),
+                target,
+                boost::numeric_cast<unsigned int>(request_.version() / 10),
+                boost::numeric_cast<unsigned int>(request_.version() % 10));
         });
 
         response_.version(request_.version());
